Fix comet detection wrapper by converting images to double
git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/mtTracker/main@9743 0bcb941b-74d1-0310-bb26-e6636707f451

+++ plusTipTracker/plusTipTrackerPackage/detectMovieComets.m
@@ -118,7 +118,7 @@ for iChan= p.ChannelIndex
     if ishandle(wtBar), waitbar(0,wtBar,logMsg); end  
     for i = p.firstFrame:p.lastFrame
         % Loading image
-        im = movieData.channels_(iChan).loadImage(i)/maxIntensity;
+        im = double(movieData.channels_(iChan).loadImage(i))/maxIntensity;
         
         % Combine region of interest with preselected segmentation output
         if ~isempty(p.MaskProcessIndex)
